graphql/requirements.txt
fastapi>=0.104.1
uvicorn
strawberry-graphql>=0.212.0
